Remove RxJava as a `compile` dependency
After making the reducers completely synchronous, it seems hard to justify the use of RxJava. A very simple Redux implementation is possible in a far more readable form without using it. I hope this lowers the barrier to entry to using something like Bansa, as well as reducing the method count of your apps if you're not using Rx for anything other than Bansa!

Also, using this implementationas as a base, it would be trivial to implement an Observable interface for state changes.

Notable Changes:

  - Introduced simple Subscription class with one method: unsubscribe.
  - Removed RxStore in favor of BaseStore
  - Change dispatch & middleware interface. These functions no longer need to return an action.
  - Update all unit tests to work with new subscriptions & middleware