BugFix: Edit Text not working in Todo List Example
There is a small bug in Anvil that causes the Edit Text view to
function incorrectly. A pull request has been filed fixing the issues:
https://github.com/zserge/anvil/pull/52, but to make the example work
until that PR has been accepted / the bug is fixed in a different way,
I’m using a locally compiled version of Anvil for the Todo List example
that solves the immediate problem.